Removed Cython - no performance improvement, refactored matt coef
